pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hippocli"
version = "0.1.0"
description = "Enterprise-grade CLI toolkit for financial data acquisition, validation, conversion, and analytics."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [{ name = "HippoCLI maintainers" }]
dependencies = [
  "typer[all]>=0.12.5",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.3.4",
  "httpx>=0.27.2",
  "tenacity>=9.0.0",
  "rich>=13.8.1",
  "PyYAML>=6.0.2",
  "pandas>=2.2.3",
  "polars>=1.12.0",
  "pyarrow>=17.0.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
hippocli = "hippocli.cli:run"

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-cov>=6.0.0",
  "ruff>=0.6.9",
  "black>=24.10.0",
  "mypy>=1.13.0",
  "types-requests",
  "pre-commit>=4.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/hippocli"]

[tool.ruff]
line-length = 100
extend-select = ["I"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_unused_ignores = true
strict_optional = true

